Allow cache write operations to overwrite existing data.
By default, cache writes always overwrite existing field data, so this may
not seem like a very consequential change.

However, when a field has a merge function defined, the merge function
will be called whenever incoming data is written to the field, receiving
as its first two parameters (0) any existing data (possibly undefined) and
(1) the incoming data.

Since merge functions generally attempt to combine incoming data with
existing data, when a merge function is defined, there is no easy way to
tell the cache to _replace_ existing field data with incoming field data,
instead of combining them.

The overwrite:true option enables the replacement behavior by causing
undefined to be passed as the existing data to any merge functions
involved in the cache write, so that the merge function behaves as if this
was the first time any data had been written for the field in question,
effectively replacing any existing data.